Add moving background.

+++ web/source/Statics.dart
@@ -9,10 +9,10 @@ class Statics {
   static int KEY_JUMP = 32;
   static int KEY_FIRE = 70;
   
-  static int TILE_SIZE = 40;
-  static int WORLD_HEIGHT = 480;
-  static int BACKGROUND_WIDTH = 800;
-  static int BACKGROUND_HEIGHT = 480;
+  static double TILE_SIZE = 40.0;
+  static double WORLD_HEIGHT = 480.0;
+  static double BACKGROUND_WIDTH = 800.0;
+  static double BACKGROUND_HEIGHT = 480.0;
   static double SPEED_X = 80.0;
   static double SPEED_Y_INITIAL = 600.0;
   static double SPEED_Y_ACCELERATE = - 30.0;
